# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nong San Xanh API",
        "description": "Tài liệu API cho dự án Nông Sản Xanh.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/products/{id}/images": {
            "post": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Upload ảnh sản phẩm",
                "description": "POST /api/v1/admin/products/{id}/images",
                "operationId": "16d6a7343b7aa0c659f50a7005ff9a7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "images[]"
                                ],
                                "properties": {
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin - upload nhiều ảnh thành công."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/products/{id}/images/reorder": {
            "patch": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Sắp xếp lại thứ tự ảnh",
                "description": "PATCH /api/v1/admin/products/{id}/images/reorder",
                "operationId": "2f2d3743351affc302225f3be2ec25a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "sort_order": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin - sắp thứ tự ảnh thành công."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Lấy giỏ hàng",
                "operationId": "702bc7d417612e9daa6f176f197155ec",
                "responses": {
                    "200": {
                        "description": "Lấy giỏ hàng hiện tại thành công."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cart/items": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Thêm sản phẩm vào giỏ",
                "operationId": "c3c03ff2d6cba6734316a76e0098332f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thêm sản phẩm vào giỏ thành công."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cart/items/{id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Xóa sản phẩm khỏi giỏ",
                "operationId": "71bdfcfe569d7cc730c51efb404e3e10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa sản phẩm khỏi giỏ thành công."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Cart"
                ],
                "summary": "Cập nhật số lượng sản phẩm trong giỏ",
                "operationId": "5d857c58e4c119b9804ce7a81d5d5c5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "number",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật số lượng trong giỏ thành công."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}